completed add certification functionality
