+++ public/edit-profile-tutor.html
@@ -183,7 +183,7 @@
                     <option selected disabled>Select a Certification</option>
                   </select>
                   <div>
-                    <button class="btn col-16" type="button">Add Certification!</button>
+                    <button id="addCertification" class="btn col-16" type="button">Add Certification!</button>
                   </div>
                   </div>
                   <div id="special-interests-header">
@@ -220,7 +220,7 @@
           </div>
         <script src="./js/choose-image.js"></script>
         <script src="./js/pullListData.js"></script>
-        <script src="./js/editTutorProfile-addInstrument.js"></script>
+        <script src="./js/editTutorProfile-addTutorDetails.js"></script>
         <script>pullInstrumentData();</script>
         <script>pullCertData();</script>
          <script>pullSpecialtyData();</script>
